DUI：初步支持自定义DPI、调整文件结构、修改部分命名、线程启动结构修改并换用可等待定时器类、拆分通用类型常量定义到单独文件；InitOA更名为InitObjAttr；CEzD2D类新增DPI参数

+++ eck/CDuiList.h
@@ -537,7 +537,7 @@ public:
 			if (m_bDraggingSel)
 			{
 				auto rcDrag = MakeD2DRcF(m_rcDragSel);
-				OffsetRect(rcDrag, 0, -m_psv->GetPos());
+				OffsetRect(rcDrag, 0.f, (float)-m_psv->GetPos());
 				auto cr = crs.crBkHot;
 				cr.a *= 0.7f;
 				m_pBrush->SetColor(cr);
